+++ pkg/server/server.go
@@ -61,4 +61,3 @@ func (s *Server) Shutdown(ctx context.Context) error {
 
 	return s.httpServer.Shutdown(ctx)
 }
-